Fix MSVC warnings created by fixing clazy warnings

+++ source/app/graph/graphmodel.cpp
@@ -224,7 +224,7 @@ void GraphModel::cancelTransformBuild()
 QStringList GraphModel::availableTransformNames() const
 {
     QStringList stringList;
-    stringList.reserve(_graphTransformFactories.size());
+    stringList.reserve(static_cast<int>(_graphTransformFactories.size()));
 
     for(auto& t : _graphTransformFactories)
     {
