+++ source/app/transform/graphtransformconfig.cpp
@@ -153,7 +153,7 @@ QVariantMap GraphTransformConfig::asVariantMap() const
     QVariantMap map;
 
     QVariantList flags;
-    flags.reserve(_flags.size());
+    flags.reserve(static_cast<int>(_flags.size()));
     for(const auto& flag : _flags)
         flags.append(flag);
     map.insert("flags", flags);
