+++ source/app/ui/visualisations/visualisationconfig.cpp
@@ -39,7 +39,7 @@ QVariantMap VisualisationConfig::asVariantMap() const
     QVariantMap map;
 
     QVariantList flags;
-    flags.reserve(_flags.size());
+    flags.reserve(static_cast<int>(_flags.size()));
     for(const auto& flag : _flags)
         flags.append(flag);
     map.insert("flags", flags);
